Update for Miniclass Website

+++ database/migrations/2021_11_03_044644_create_mahasiswa_matakuliah_table.php
@@ -15,7 +15,7 @@ class CreateMahasiswaMatakuliahTable extends Migration
     {
         Schema::create('mahasiswa_matakuliah', function (Blueprint $table) {
             $table->increments('id_nilai');
-            $table->unsignedInteger('mahasiswa_id')->nullable();
+            $table->unsignedBigInteger('mahasiswa_id')->nullable();
             $table->bigInteger('matakuliah_id')->unsigned()->nullable();
             $table->string('nilai', 7)->nullable();
             $table->timestamps();
